Create Imagetotext.py
You can use the PIL module in sync with the pytesseract module to read literal images and translate it into text. 
From doing this I learned that for line 24 if you write r"\n", it doesn't read it and skips it. To delete new lines, you need to put it in a f string not a raw string. That's a werid interaction...